% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muschelli_description.R
\name{muschelli_description}
\alias{muschelli_description}
\title{DESCRIPTION for a Muschelli Package}
\usage{
muschelli_description(fields = muschelli_fields())
}
\arguments{
\item{fields}{A named list of fields to add to "DESCRIPTION",
passed to \code{\link{use_description}}}
}
\description{
DESCRIPTION for a Muschelli Package
}
