% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muschelli_rproj.R
\name{muschelli_rproj}
\alias{muschelli_rproj}
\title{Use Muschelli RProject file}
\usage{
muschelli_rproj(base_path = ".")
}
\arguments{
\item{base_path}{Path to package root.}
}
\value{
Invisible TRUE
}
\description{
Use Muschelli RProject file
}
