% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muschelli_appveyor.R
\name{use_muschelli_appveyor}
\alias{use_muschelli_appveyor}
\title{Use Muschelli Travis file}
\usage{
use_muschelli_appveyor(base_path = ".")
}
\arguments{
\item{base_path}{Path to package root.}
}
\value{
Invisible TRUE
}
\description{
Use Muschelli Travis file
}
