% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muschelli_fields.R
\name{muschelli_fields}
\alias{muschelli_fields}
\title{Fields for a Muschelli Package}
\usage{
muschelli_fields(title = "", description = "", given = "John",
  family = "Muschelli", email = "muschellij2@gmail.com",
  role = c("aut", "cre"), comment = c(ORCID = "0000-0001-6469-1750"))
}
\arguments{
\item{title}{Title of package}

\item{description}{Description of package.  Should be able to work with the
start of the sentence "The goal is to provide".}

\item{given}{a character vector with the given names, or a list thereof.}

\item{family}{a character string with the family name, or a list thereof.}

\item{email}{a character string (or vector) giving an e-mail address
(each), or a list thereof.}

\item{role}{Role of the people (author/creator)}

\item{comment}{Additional comments (usually for ORCID ID)}
}
\value{
A list of named fields
}
\description{
Fields for a Muschelli Package
}
\examples{
muschelli_fields()
}
