% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badges.R
\name{travis_badge}
\alias{travis_badge}
\alias{appveyor_badge}
\alias{coverage_badge}
\alias{coveralls_badge}
\alias{codecov_badge}
\title{Travis Badge}
\usage{
travis_badge(gh_username)

appveyor_badge(gh_username)

coverage_badge(..., coverage_type = "codecov")

coveralls_badge(gh_username)

codecov_badge(gh_username)
}
\arguments{
\item{gh_username}{GitHub username}

\item{...}{not used}

\item{coverage_type}{CI tool to use.
Currently supports codecov and coveralls.}
}
\value{
Character Vector
}
\description{
Travis Badge
}
