% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muschelli_rmd.R
\name{use_muschelli_readme_rmd}
\alias{use_muschelli_readme_rmd}
\title{Muschelli README Rmd File}
\usage{
use_muschelli_readme_rmd(coverage_type = "codecov")
}
\arguments{
\item{coverage_type}{CI tool to use.
Currently supports codecov and coveralls.}
}
\value{
Invisible TRUE
}
\description{
Muschelli README Rmd File
}
