% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muschelli_travis.R
\name{use_muschelli_travis}
\alias{use_muschelli_travis}
\title{Use Muschelli Travis file}
\usage{
use_muschelli_travis(coverage_type = "codecov")
}
\arguments{
\item{coverage_type}{CI tool to use.
Currently supports codecov and coveralls.}
}
\value{
Invisible TRUE
}
\description{
Use Muschelli Travis file
}
